+++ src/types/index.ts
@@ -1,3 +1,5 @@
+export type Layout = 'main' | 'favorites' | 'other';
+
 export type Offer = {
   id: string;
   title: string;
@@ -43,3 +45,19 @@ export type OfferHost = {
 export type OfferType = 'Apartment' | 'Private room';
 
 export type Offers = Offer[];
+
+export type Comment = {
+  id: string;
+  date: string;
+  user: CommentUser;
+  comment: string;
+  rating: number;
+};
+
+export type CommentUser = {
+  name: string;
+  avatarUrl: string;
+  isPro: boolean;
+};
+
+export type Comments = Comment[];
