--- src/types/layout.ts
@@ -1,1 +1,0 @@
-export type Layout = 'main' | 'favorites' | 'other';
